Added second song
There are many separate variables for 2 levels.. If I'm going to cleanup GameManager, I'll probably make a struct to store data for levels separately

+++ Assets/Scripts/ScoreUIGrabber.cs
@@ -9,7 +9,7 @@ public class ScoreUIGrabber : MonoBehaviour
     int miss;
     int hit;
     int perfectHit;
-    int ratingScore;
+    float ratingScore;
 
     public TextMeshProUGUI missText;
     public TextMeshProUGUI hitText;
